fix a few things

+++ Events/RegionManager.cs
@@ -7,10 +7,11 @@ using BattleBitMinigames.Handlers;
 using BattleBitMinigames.Interfaces;
 
 namespace BattleBitMinigames.Events;
+
 public class RegionManager : Event
 {
     private CancellationTokenSource? _cancellationTokenSource;
-    
+
     private void StartRegionManager()
     {
         Program.Logger.Info("Started RegionManager!");
@@ -21,47 +22,61 @@ public class RegionManager : Event
         {
             while (Server.IsConnected && !cancellationToken.IsCancellationRequested)
             {
-                foreach (var player in Server.AllPlayers.Where(player => player.IsAlive && player.Position != Vector3.Zero))
+                foreach (var player in Server.AllPlayers.Where(player =>
+                             player.IsAlive && player.Position != Vector3.Zero))
                 {
                     var region = RegionHelper.GetIsPlayerInRegion(RegionList.GetMapRegions(Server.Map), player);
                     if (region != null)
                     {
-                        var now = DateTime.UtcNow;
-                        var spawnedInSpawn = player.GetPlayerProperty(PlayerProperties.IVipPlayerProperties.SpawnedInSpawn);
-                        var spawnedInSpawnTimeStr = player.GetPlayerProperty(PlayerProperties.IVipPlayerProperties.SpawnedInSpawnTime);
-                        var enteredSpawnTimeStr = player.GetPlayerProperty(PlayerProperties.IVipPlayerProperties.EnteredSpawnTime);
+                        if (Program.ServerConfiguration.LaunchCustomGamemode == "vip")
+                        {
+                            var isVip = player.GetPlayerProperty(PlayerProperties.IVipPlayerProperties.IsVip);
+                            if (isVip == "" || isVip == "false") continue;
 
-                        if (spawnedInSpawn == "true")
-                        {
-                            if (DateTime.TryParse(spawnedInSpawnTimeStr, out var spawnedInSpawnTime))
+                            var now = DateTime.UtcNow;
+                            var spawnedInSpawn =
+                                player.GetPlayerProperty(PlayerProperties.IVipPlayerProperties.SpawnedInSpawn);
+                            var spawnedInSpawnTimeStr =
+                                player.GetPlayerProperty(PlayerProperties.IVipPlayerProperties.SpawnedInSpawnTime);
+                            var enteredSpawnTimeStr =
+                                player.GetPlayerProperty(PlayerProperties.IVipPlayerProperties.EnteredSpawnTime);
+
+                            if (spawnedInSpawn == "true")
                             {
-                                if ((now - spawnedInSpawnTime).TotalSeconds >= 90)
+                                if (DateTime.TryParse(spawnedInSpawnTimeStr, out var spawnedInSpawnTime))
                                 {
-                                    player.Kill();
-                                    player.Message($"You spawned in the {region.Name} and stayed for too long!", 15);
-                                }
-                                else
-                                {
-                                    player.Message($"You spawned in the {region.Name}, please leave within {(int)(90 - (now - spawnedInSpawnTime).TotalSeconds)} seconds!");
+                                    if ((now - spawnedInSpawnTime).TotalSeconds >= 90)
+                                    {
+                                        player.Kill();
+                                        player.Message($"You spawned in the {region.Name} and stayed for too long!",
+                                            15);
+                                    }
+                                    else
+                                    {
+                                        player.Message(
+                                            $"You spawned in the {region.Name}, please leave within {(int)(90 - (now - spawnedInSpawnTime).TotalSeconds)} seconds!");
+                                    }
                                 }
                             }
-                        }
-                        else
-                        {
-                            if (enteredSpawnTimeStr == string.Empty)
+                            else
                             {
-                                player.SetPlayerProperty(PlayerProperties.IVipPlayerProperties.EnteredSpawnTime, now.ToUniversalTime().ToString());
-                            }
-                            else if (DateTime.TryParse(enteredSpawnTimeStr, out var enteredSpawnTime))
-                            {
-                                if ((now - enteredSpawnTime).TotalSeconds >= 20)
+                                if (enteredSpawnTimeStr == string.Empty)
                                 {
-                                    player.Kill();
-                                    player.Message($"You were in the {region.Name} for too long!", 15);
+                                    player.SetPlayerProperty(PlayerProperties.IVipPlayerProperties.EnteredSpawnTime,
+                                        now.ToUniversalTime().ToString());
                                 }
-                                else
+                                else if (DateTime.TryParse(enteredSpawnTimeStr, out var enteredSpawnTime))
                                 {
-                                    player.Message($"You entered the {region.Name}, please leave within {(int)(20 - (now - enteredSpawnTime).TotalSeconds)} seconds!");
+                                    if ((now - enteredSpawnTime).TotalSeconds >= 20)
+                                    {
+                                        player.Kill();
+                                        player.Message($"You were in the {region.Name} for too long!", 15);
+                                    }
+                                    else
+                                    {
+                                        player.Message(
+                                            $"You entered the {region.Name}, please leave within {(int)(20 - (now - enteredSpawnTime).TotalSeconds)} seconds!");
+                                    }
                                 }
                             }
                         }
@@ -77,7 +92,7 @@ public class RegionManager : Event
 
                 await Task.Delay(1000, cancellationToken);
             }
-            
+
             Program.Logger.Info("RegionManager stopped.");
         }, cancellationToken);
     }
@@ -104,22 +119,23 @@ public class RegionManager : Event
                 StopRegionManager();
             }
         }
-        
+
         return base.OnPlayerTypedMessage(player, channel, msg);
     }
-    
+
     public override Task OnPlayerSpawned(BattleBitPlayer player)
     {
         var region = RegionHelper.GetIsPlayerInRegion(RegionList.GetMapRegions(Server.Map), player);
         if (region != null)
         {
             player.SetPlayerProperty(PlayerProperties.IVipPlayerProperties.SpawnedInSpawn, "true");
-            player.SetPlayerProperty(PlayerProperties.IVipPlayerProperties.SpawnedInSpawnTime, DateTime.UtcNow.ToString());
+            player.SetPlayerProperty(PlayerProperties.IVipPlayerProperties.SpawnedInSpawnTime,
+                DateTime.UtcNow.ToString());
         }
-        
+
         return base.OnPlayerSpawned(player);
     }
-    
+
     public override Task OnConnected()
     {
         StartRegionManager();
